rename allocator package to subnetallocator in prep for ipallocator pkg
