+++ go-controller/pkg/ovn/subnetallocator/allocator.go
@@ -1,4 +1,4 @@
-package allocator
+package subnetallocator
 
 import (
 	"fmt"
